# docs/swagger.yaml
basePath: /
definitions:
  models.BabyName:
    properties:
      birth_year:
        type: integer
      child_first_name:
        type: string
      count:
        type: integer
      ethnicity:
        type: string
      gender:
        type: string
      id:
        type: string
      rank:
        type: integer
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for baby names API.
  title: Baby Names API
  version: "1.0"
paths:
  /api/babynames:
    get:
      consumes:
      - application/json
      description: Get all baby names
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BabyName'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all baby names
      tags:
      - babynames
  /api/babynames/ethnicity:
    get:
      consumes:
      - application/json
      description: Get baby names by ethnicity
      parameters:
      - description: Ethnicity
        in: query
        name: ethnicity
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BabyName'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get baby names by ethnicity
      tags:
      - babynames
  /api/babynames/gender:
    get:
      consumes:
      - application/json
      description: Get baby names by gender
      parameters:
      - description: Gender
        in: query
        name: gender
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BabyName'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get baby names by gender
      tags:
      - babynames
  /api/babynames/id:
    get:
      consumes:
      - application/json
      description: Get baby name by ID
      parameters:
      - description: BabyName ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BabyName'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get baby name by ID
      tags:
      - babynames
  /api/babynames/year:
    get:
      consumes:
      - application/json
      description: Get baby names by birth year
      parameters:
      - description: Birth Year
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BabyName'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get baby names by birth year
      tags:
      - babynames
swagger: "2.0"